Change package name of JAVA STREAMS
