+++ src/main/java/com/enh/java/streams/enums/Arrival.java
@@ -1,0 +1,5 @@
+package com.enh.java.streams.enums;
+
+public enum Arrival {
+    NEW, OLD;
+}
