+++ src/main/java/com/enh/java/streams/enums/Brand.java
@@ -1,0 +1,5 @@
+package com.enh.java.streams.enums;
+
+public enum Brand {
+    SAMSUNG, APPLE, ONE_PLUS, XIAOMI;
+}
