+++ src/main/java/com/enh/java/streams/enums/PaymentMode.java
@@ -1,0 +1,5 @@
+package com.enh.java.streams.enums;
+
+public enum PaymentMode {
+    PAYPAL, ADYEN, CREDIT_CARD, DEBIT_CARD, AMAZON_PAY, PAYTM;
+}
