+++ src/main/java/com/enh/java/streams/enums/Role.java
@@ -1,0 +1,5 @@
+package com.enh.java.streams.enums;
+
+public enum Role {
+    SAP_BACKEND_DEV, SAP_FRONTEND_DEV, SAP_QA_TESTER, SAP_COMMERCE_TECH_LEAD, SAP_COMMERCE_ARCHITECT, SAP_COMMERCE_DELIVERY_HEAD;
+}
