--- src/main/java/com/enh/streams/enums/Arrival.java
@@ -1,5 +1,0 @@
-package com.enh.streams.enums;
-
-public enum Arrival {
-    NEW, OLD;
-}
