--- src/main/java/com/enh/streams/enums/Brand.java
@@ -1,5 +1,0 @@
-package com.enh.streams.enums;
-
-public enum Brand {
-    SAMSUNG, APPLE, ONE_PLUS, XIAOMI;
-}
