--- src/main/java/com/enh/streams/enums/Department.java
@@ -1,5 +1,0 @@
-package com.enh.streams.enums;
-
-public enum Department {
-    BE_DEVELOPEMENT, FE_DEVELOPMENT, TECH_DESIGN, FEATURE_TESTING, TESTING, HR, FINANCIAL, ADMINISTRATOR;
-}
