--- src/main/java/com/enh/streams/enums/Gender.java
@@ -1,5 +1,0 @@
-package com.enh.streams.enums;
-
-public enum Gender {
-    MALE, FEMALE;
-}
