--- src/main/java/com/enh/streams/enums/PaymentMode.java
@@ -1,5 +1,0 @@
-package com.enh.streams.enums;
-
-public enum PaymentMode {
-    PAYPAL, ADYEN, CREDIT_CARD, DEBIT_CARD, AMAZON_PAY, PAYTM;
-}
